Add a generic Provider interface
